s21-day06/wip -> ex01 : add architecture
