+++ day06/ex01/internal/admin_credentials.txt
@@ -1,4 +1,4 @@
 adminLogin:zatannal
 adminPassword:zatannal
-dbName:kek
-dbTable:lol
+dbName:postgres
+dbUser:postgres
